Recipe and bb bug fix of electric fence gate

Approved-by: Poopoo Dice <poopoodice2288@gmail.com>

+++ src/main/java/net/kaneka/planttech2/items/FertilizerItem.java
@@ -82,7 +82,7 @@ public class FertilizerItem extends BaseItem
     @Override
     public void addInformation(ItemStack stack, World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn)
     {
-	tooltip.add(new StringTextComponent(new TranslationTextComponent("info.fertilizer").getUnformattedComponentText() + ": " + (getIncreaseChance(stack.getItem()) * 100) + "%"));
+	tooltip.add(new StringTextComponent(new TranslationTextComponent("info.fertilizer").getUnformattedComponentText() + ": " + (int) (getIncreaseChance(stack.getItem()) * 100) + "%"));
 	if(stack.getItem() == ModItems.FERTILIZER_CREATIVE)
 	{
 	    tooltip.add(new StringTextComponent(new TranslationTextComponent("info.fertilizer_creative").getUnformattedComponentText()));
